Check WordPress and PHP requirements before installing a theme or plugin
This matches the behavior of WordPress core which will refuse to install
a plugin if the local copy of WordPress or PHP don't meet the minimum
requirements for the most recent version as listed by the plugin
authors.

Unfortunately the api is limited and only provides these requirement
details for the most recent version, so it isn't possible to find an
older version that might work. As a compromise, this code doesn't check
requirements if a user provides a specific --version since we can't know
the requirements for anything other than the latest version and assume
if somebody specifies a version they know it will work or want to try
anyway.