More updates to work on WordPress 4.9
Almost all of these have the same cause -- WordPress ships with Akismet
but the recent versions do not work on WordPress 4.9 so commands like wp
plugin list now show the updates as unavailable and with two additional
table headers for requires and requires_php that explain why. So just be
intentional about the fields we care about testing.